server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
